Late Colonial
1660-1763

As time progressed, slavery became more and more crucial to the economy of the English colonies, as they moved away from indentured servitude. This was in part due to Bacon’s Rebellion, which made the white elite fear an uprising among their workers. It was also due to the higher life expectancy, which made buying a worker for life became more profitable. Slaves’ freedom was hurt even further by the tightening of slave codes, a result of the Stono Rebellion in 1739, as well as the slave uprising that occurred in New York. Enslaved women were even less free, as they often had to deal with a court system that refused to acknowledge the sexual abuse commited by their masters. Slavery would also give way to a more oppressive and rigid class structure, with those owning many slaves comprising the upper echelon of the elite, and those with no slaves being impoverished.

The English would continue to treat Indians more and more subhuman, as the many conflicts with them, such as Metacom’s Rebellion, made the colonists think them as barbaric and savage. Bacon’s Rebellion was fought in part because Lord Berkely did not treat the Indians with the rebels' desired level of scorn. In truth, the barbaric savages were the colonists themselves, who thought it to be their “freedom” to burn Indian villages and steal their land. This “freedom” was altered by the Proclamation of 1763, which restricted settlement beyond the Appalachians, and angered the colonists.

The authority of the government in matters of religion also began to be questioned during this time. The Salem Witch trials, which took place between February 1692 and May 1963, showed the theocracy that existed in the Puritan colonies (and witch trials in general) to be a joke, as it proved to be one of the most exemplary cases of religious hysteria, with many being executed without any real reason. The Great Awakening likewise questioned theocracy, as well as the authority of the government in general. Influenced by the Enlightenment, it called for religious fundamentalism and freedom, and would shape the coming revolution, as well as democratic thought in general. The debate over English freedom that started with the English Civil War also expanded the definition of freedom such that the king could not simply rule with an iron fist, but was required to preserve certain rights for his subjects.  Finally, the chains around people’s freedom of expression began to loosen, thanks in part to the John Peter Zenger trial, in which it was argued that truth was a defense against libel.