Revolutionary
1763-1783

The American Revolution was a revolution of ideas. While the concept of universal rights and freedom for all had existed to an extent before, it was the Revolution that brought it to the forefront in America. In the Declaration of Independence, Thomas Jefferson stated that “all men are created equal”, which was a radical departure from the ideas that had so dominated England for centuries. It also made many question the monarchical government of England. In Thomas Paine’s Common Sense, one of the most important documents of the Revolution, he stated that “There is something exceedingly ridiculous in the composition of Monarchy; it first excludes a man from the means of information, yet empowers him to act in cases where the highest judgment is required.”

However, ideas are not reality, and in reality, the Revolution did not change the freedom that many oppressed groups had. The most obvious contradiction to Jefferson’s famous dictum is that slavery continued to exist well after the revolution, and it was not one that went unrecognized. Jefferson himself noted that “The whole commerce between master and slave is a perpetual exercise of the most boisterous passions, the most unremitting despotism on the one part, and degrading submissions on the other.” Not only did the horrific institution of slavery continue to degrade the freedom of blacks, but there were no long term changes in women’s freedom. Although, the concept of republican motherhood did emerge in this era, and though women were still limited to their families, it still promoted women’s education and made their work more important.

Note: I removed the “Native Americans” section not because they ceased to exist, but because there aren’t many notable changes in their own conception of freedom, which is what that section was dedicated to.