Federalist
1783-1801

Immediately following the Revolution, there was a general mistrust of centralized government. For this reason, under the Articles of Confederation, states had complete freedom, with very little power granted to the national government. Upon realizing that there is such a thing as too much liberty, the Constitution was drafted, in which there was a centralized government, but with many checks and balances to ensure no one branch poses a threat to liberty. The Constitution represented a shift from viewing centralized government as a threat to individuals and their liberty, to instead viewing it as a means of promoting democracy.

There was much disagreement on the new republic should be ran. On one hand were the federalists such as Hamilton, who sought to take the good parts of Britain’s government, such as its central bank, and leave out the bad parts. Hamilton revealed a 5 step economic plan involving a tax on  Many saw this as an affront to freedom, such as those involved in the Whiskey Rebellion, who echoed the demands of the Revolution. The government also limited freedom of speech with the Alien and Sedition acts, which authorized those who criticized the government to be prosecuted.

While there was no real progress made in terms of government legislation, women gained an upper hand in the realm of ideas. Mary Wollstonecraft, for one, published her pamphlet titled “A Vindication of the Rights of Women”, directly challenging the contradiction that “universal” rights were being limited to males. Judith Sargent Murray wrote her own essay titled “On the Equality of the Sexes”, which insisted that “our [women’s] souls are by nature equal to yours; the same breath of God animates, enlivens, and invigorates us.” Both of these women helped lay the seeds of the future feminist movement.