Jeffersonian
1801-1815

Opposite of the Federalists were the Republicans, who thought that politics was not the job of the elite, but of ordinary people. The face of the Republican party, Thomas Jefferson, was elected in 1801, and one would think that his policies would reflect his desire for limited government. Rather, he used his executive power to great effect, using it to purchase the sprawling Louisiana Territory, which destroyed the freedom of the many blacks, Indians, and women that had already been living there.

Following the invention of the cotton gin by Eli Whitney, the institution of slavery was expanded substantially. What was previously an unprofitable crop to farm now became the most profitable, and dominated the entire economy of the South, further tying it to slavery. Slaves were given a new hope but also had their freedom further hurt by the slave rebellions that were occurring at the time. Both the first and largest slave rebellion, the Haitian Revolution, was the match that lit the fuse, and frightened slaveholders worldwide. Gabriel’s Rebellion took place in the wake of the Haitian Revolution, and while it would hurt the institution of slavery in the long term, it led to a tightening of slave laws.