Jacksonian
1815-1840

The Market Revolution resulted in a change in gender norms, as many women were no longer confined entirely to their homes, and instead worked factory jobs. The textile mills of Lowell, Massachusetts had a workforce that was dominated by young women. Many of these jobs were short lived, however, and a new perception of gender norms dubbed the “cult of domesticity” arose. Essentially a degradation of the earlier idea of republican motherhood, it not only thought female frailty, innocence, and weakness to be a fact of life, but romanticized these traits, many believing them to be virtuous.

Jacksonian democracy was essentially a drastic growth of the ideas of Jeffersonian democracy, in that it sought to grant freedom and a say in political matters to not only those with wealth, but also to the everyman, those in the lower class. During this time, the property requirement for voting was completely eliminated, allowing any white men to vote. That status of being white and male, however, was as important in this period as ever, as all other groups did not have greater freedom, or in some cases, were less free than before. Noted Indian massacrer Jackson treated the Native Americans with scorn, and passed the Indian Removal Act in 1830. It moved many Indians to lands that were completely unfamiliar to them, and were, as stated in the Appeal of the Cherokee Nation, “badly supplied with wood and water; and no Indian tribe can live as agriculturists without these articles.” The one chance that Indians had at being legally recognized as a sovereign people, Worcester v. Georgia, was struck down by Jackson’s mighty executive hand. Quite like Jefferson, Jackson’s advocacy for states rights and a limited national government was contradicted by his use of executive power. In the Nullification Crisis, for instance, Jackson used military force against South Carolina because they tried to nullify a tariff.

All in all, the period of Jackson’s presidency was one of unprecedented levels of freedom for white men, but little to no freedom for anyone else.