Age of Reform
1820-1840

While slavery had previously been justified by claiming it be a necessary evil, many now believed that it was somehow good for its subjects. John C. Calhoun was one of these people, as he believed the condition of the enslaved to be far preferable to that of the “tenants of the poor houses in the more civilized portions of Europe.” There was also the paternalist ethos that had existed since the early 1800s, which made many view slaveholders as the protectors of their slaves. The institution was also weakened, however, by the many fugitive slaves that escaped their conditions, showing that many slaves do have aspirations beyond it. The Underground Railroad, organized by Hariet Tubman, was a decentralized abolitionist network that is said to have freed around 100,000 slaves. Many of these free blacks had their freedom limited by the governments of the Southern states, however, which often expelled them or prohibited them from entering.

During this time, there was also a growing abolitionist movement, which altered the ideas surrounding freedom for blacks. William Loyd Garrison was a key figure in this movement, as he published the abolitionist newspaper titled “The Liberator”, which spread throughout the South due to his many critics discussing it. Literature also played a role in the abolitionist movement. “Uncle Tom’s Cabin”, written by Harriet Beecher Stowe, was effective in convincing many of the evils of slavery, as it portrayed the slaves as sympathetic and feeling while showing the masters to be cruel and malicious.

This period saw an opening of the public sphere to the voices of women. Dorothea Dix was one of these women, who advocated for the creation of asylums and for more humane treatment of the mentally ill. The public sphere was also open to women who wanted to voice concern for their own freedom. The Seneca Falls Convention, organized primarily by Elizabeth Cady Stanton and Lucretia Mott, resulted in the Declaration of Sentiments, which stated that “The history of mankind is a history of repeated injuries and usurpations on the part of man toward woman.” The Seneca Falls Convention also marked the first significant discussion of women’s suffrage. Women also played a major role in the abolitionist movement, the Grimké Sisters being the first to apply the universal freedom of abolitionism to the subject of women’s rights.