Civil War
1840-1865

Prior to the Civil War, the question of slavery continued to become the most debated subject. The Mexican-American War amplified these conflicts, as there was much debate as to whether the new territories should be slave states. The Kansas-Nebraska Act merely amplified these conflicts, as it effectively repealed the Missouri Compromise, and led to a near civil war in Kansas.

The freedom of blacks was further worsened by the federal government. The Dred Scott case, which occurred in 1857, came to the insane conclusion that blacks were not even meant to be included as citizens of the United States. The Fugitive Slave Law was also passed during this time, and it rewarded those who captured escaped slaves monetarily, and penalized those who failed to do so. Both of these events furthered the fear that the federal government was now ran by slaveholders. They did much to move many to the side of abolitionism.

The Civil War began on April 12, 1861, when the Confederates attacked Fort Sumter. It was a direct clash between the institution of slavery and the notion of all men being created equal that the United States was founded on. In the South, this was clearly the case from the very beginning, considering every one of the states seceded in order to protect slavery, them perceiving an attack on slavery to be an attack on the South. This wasn’t the immediate case in the North, however, as many Northerners, such as Lincoln himself, insisted the fight was not centered around slavery and was instead about preserving the union. This would inevitably cease to be the case, as the war had merged this desire for unity with the desire to exterminate slavery, and as said by Frederick Douglass, the two causes had “become identical”. Lincoln would issue the Emancipation Proclamation in 1862, which stated that in states not under control of the Union, slaves “henceforward shall be free”. The new goal of the Civil War to redefine freedom to also include African Americans was further solidified in Lincoln’s Gettysburg Address, which not only stated the meaning of the war to be preservation of the Union, but also stated that it would lead to “a new birth of freedom”. Slavery would not be completely abolished until after the war ended, with the passage of the 13th amendment, which was also the first direct mention of slavery in the constitution.